server/requirements.txt
fastapi>=0.135.0
uvicorn>=0.44.0
pyyaml>=6.0
mem0ai>=1.0.11
chromadb>=1.5.0
sentence-transformers>=5.0.0
google-genai>=1.70.0
apscheduler>=3.10.0
